requirements.txt
streamlit
pandas
python-dotenv
snowflake-snowpark-python[pandas]
